modularisation partie 1
